Ajustes na classe Curso e em suas devidas pastas, atualização na classe Instrutor e suas devidas pastas

+++ plataformacursos/plataformacursos/src/main/java/com/veridia/gestao/plataformacursos/Service/CursoService.java
@@ -1,17 +1,20 @@
-package com.veridia.plataforma.service;
+package com.veridia.gestao.plataformacursos.Service;
 
-import com.veridia.plataforma.domain.model.Curso;
-import com.veridia.plataforma.domain.repository.CursoRepository;
+import com.veridia.gestao.plataformacursos.Curso;
+import com.veridia.gestao.plataformacursos.Repository.CursoRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
 
+
 @Service
 public class CursoService {
 
-    @Autowired // Injeta o Repositório para acessar o banco de dados
+    @Autowired
     private CursoRepository cursoRepository;
 
     public List<Curso> buscarTodos() {
@@ -22,17 +25,29 @@ public class CursoService {
         return cursoRepository.findById(id);
     }
 
+    @Transactional
     public Curso salvar(Curso curso) {
-        // **AQUI ENTRA A LÓGICA DE NEGÓCIO, EXEMPLO:**
-        // 1. Validar se o preço é positivo.
-        // 2. Garantir que o Instrutor existe, etc.
-
         return cursoRepository.save(curso);
     }
 
+    @Transactional
+    public Curso atualizar(Long id, Curso cursoAtualizado) {
+        Curso curso = cursoRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Curso não encontrado com ID: " + id));
+
+        curso.setNome(cursoAtualizado.getNome());
+        curso.setDescricao(cursoAtualizado.getDescricao());
+        curso.setPreco(cursoAtualizado.getPreco());
+        curso.setLimiteVagas(cursoAtualizado.getLimiteVagas());
+        curso.setInstrutor(cursoAtualizado.getInstrutor());
+        return cursoRepository.save(curso);
+    }
+
+    @Transactional
     public void deletar(Long id) {
-        // **AQUI ENTRA A LÓGICA DE NEGÓCIO, EXEMPLO:**
-        // 1. Verificar se há alunos inscritos antes de deletar.
+        if (!cursoRepository.existsById(id)) {
+            throw new EntityNotFoundException("Curso não encontrado com ID: " + id);
+        }
         cursoRepository.deleteById(id);
     }
 }
